optimize发布视频：线上要把视频先存储在本地，才能够去获取封面图

+++ laravel-api/app/Models/Dynamic/Topic.php
@@ -103,6 +103,6 @@ class Topic extends Model
      */
     public static function getDetaultTopicId(): int
     {
-        return self::where('is_default', 1)->value('topic_id', 0);
+        return (int)self::where('is_default', 1)->value('topic_id', 0);
     }
 }
